requirements.txt
openai==1.3.5
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
black==23.9.1
flake8==6.1.0